requirements.txt
amqp==2.5.1
asn1crypto==1.0.0
billiard==3.6.1.0
braintree==3.56.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.5
django-cors-headers==3.1.1
django-countries==5.5
django-filter==2.2.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
djangorestframework-filters==1.0.0.dev0
entrypoints==0.3
flake8==3.7.8
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
kombu==4.6.4
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wincertstore==0.2
zipp==0.6.0
